# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MealPlanMealType = typeof MealPlanMealType[keyof typeof MealPlanMealType];


export const MealPlanMealType = {
  dinner: 'dinner',
  lunch: 'lunch',
  breakfast: 'breakfast',
} as const;

export type MealPlanStatus = typeof MealPlanStatus[keyof typeof MealPlanStatus];


export const MealPlanStatus = {
  planned: 'planned',
  cooked: 'cooked',
  skipped: 'skipped',
} as const;

export interface MealPlan {
  id: number;
  date: string;
  dayLabel: string;
  title: string;
  mealType: MealPlanMealType;
  status: MealPlanStatus;
  /** @nullable */
  recipeId?: number | null;
  /** @nullable */
  recipeName?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface ShoppingPreviewItem {
  id: number;
  name: string;
  category: string;
  completed: boolean;
}

export type NurseryStatusTone = typeof NurseryStatusTone[keyof typeof NurseryStatusTone];


export const NurseryStatusTone = {
  ok: 'ok',
  attention: 'attention',
  neutral: 'neutral',
} as const;

export interface NurseryStatus {
  label: string;
  detail: string;
  tone: NurseryStatusTone;
}

export interface DashboardSummary {
  dateLabel: string;
  weekLabel: string;
  todayMeal: MealPlan | null;
  weekMenu: MealPlan[];
  shoppingPreview: ShoppingPreviewItem[];
  nurseryStatus: NurseryStatus;
}

export type MealPlanInputMealType = typeof MealPlanInputMealType[keyof typeof MealPlanInputMealType];


export const MealPlanInputMealType = {
  dinner: 'dinner',
  lunch: 'lunch',
  breakfast: 'breakfast',
} as const;

export type MealPlanInputStatus = typeof MealPlanInputStatus[keyof typeof MealPlanInputStatus];


export const MealPlanInputStatus = {
  planned: 'planned',
  cooked: 'cooked',
  skipped: 'skipped',
} as const;

export interface MealPlanInput {
  date: string;
  /** @minLength 1 */
  title: string;
  mealType: MealPlanInputMealType;
  status?: MealPlanInputStatus;
  /** @nullable */
  recipeId?: number | null;
  /** @nullable */
  note?: string | null;
}

export type MealPlanUpdateMealType = typeof MealPlanUpdateMealType[keyof typeof MealPlanUpdateMealType];


export const MealPlanUpdateMealType = {
  dinner: 'dinner',
  lunch: 'lunch',
  breakfast: 'breakfast',
} as const;

export type MealPlanUpdateStatus = typeof MealPlanUpdateStatus[keyof typeof MealPlanUpdateStatus];


export const MealPlanUpdateStatus = {
  planned: 'planned',
  cooked: 'cooked',
  skipped: 'skipped',
} as const;

export interface MealPlanUpdate {
  date?: string;
  /** @minLength 1 */
  title?: string;
  mealType?: MealPlanUpdateMealType;
  status?: MealPlanUpdateStatus;
  /** @nullable */
  recipeId?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface Recipe {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  category: string;
  prepMinutes: number;
  servings: number;
  ingredients: string[];
  steps: string[];
  favorite: boolean;
}

export interface RecipeInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @minimum 1 */
  prepMinutes: number;
  /** @minimum 1 */
  servings: number;
  ingredients: string[];
  steps: string[];
  favorite?: boolean;
}

export interface RecipeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  /** @minimum 1 */
  prepMinutes?: number;
  /** @minimum 1 */
  servings?: number;
  ingredients?: string[];
  steps?: string[];
  favorite?: boolean;
}

export type ListMealPlansParams = {
weekStart?: string;
};

export type ListRecipesParams = {
search?: string;
category?: string;
};