# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Home dashboard summaries
  - name: meal-plans
    description: Weekly meal planning
  - name: recipes
    description: Recipe management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get the home dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /meal-plans:
    get:
      operationId: listMealPlans
      tags: [meal-plans]
      summary: List meal plans for a week
      parameters:
        - name: weekStart
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meal plan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlan"
    post:
      operationId: createMealPlan
      tags: [meal-plans]
      summary: Add a meal to the weekly plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanInput"
      responses:
        "201":
          description: Created meal plan entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"
  /meal-plans/{id}:
    patch:
      operationId: updateMealPlan
      tags: [meal-plans]
      summary: Update a planned meal
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanUpdate"
      responses:
        "200":
          description: Updated meal plan entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"
        "404":
          description: Meal plan not found
    delete:
      operationId: deleteMealPlan
      tags: [meal-plans]
      summary: Remove a planned meal
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Meal plan removed
        "404":
          description: Meal plan not found
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List recipes
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeUpdate"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Recipe deleted
        "404":
          description: Recipe not found
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DashboardSummary:
      type: object
      required: [dateLabel, weekLabel, todayMeal, weekMenu, shoppingPreview, nurseryStatus]
      properties:
        dateLabel:
          type: string
        weekLabel:
          type: string
        todayMeal:
          anyOf:
            - $ref: "#/components/schemas/MealPlan"
            - type: "null"
        weekMenu:
          type: array
          items:
            $ref: "#/components/schemas/MealPlan"
        shoppingPreview:
          type: array
          items:
            $ref: "#/components/schemas/ShoppingPreviewItem"
        nurseryStatus:
          $ref: "#/components/schemas/NurseryStatus"
    ShoppingPreviewItem:
      type: object
      required: [id, name, category, completed]
      properties:
        id:
          type: number
        name:
          type: string
        category:
          type: string
        completed:
          type: boolean
    NurseryStatus:
      type: object
      required: [label, detail, tone]
      properties:
        label:
          type: string
        detail:
          type: string
        tone:
          type: string
          enum: [ok, attention, neutral]
    MealPlan:
      type: object
      required: [id, date, dayLabel, title, mealType, status]
      properties:
        id:
          type: number
        date:
          type: string
          format: date
        dayLabel:
          type: string
        title:
          type: string
        mealType:
          type: string
          enum: [dinner, lunch, breakfast]
        status:
          type: string
          enum: [planned, cooked, skipped]
        recipeId:
          type: [number, "null"]
        recipeName:
          type: [string, "null"]
        note:
          type: [string, "null"]
    MealPlanInput:
      type: object
      required: [date, title, mealType]
      properties:
        date:
          type: string
          format: date
        title:
          type: string
          minLength: 1
        mealType:
          type: string
          enum: [dinner, lunch, breakfast]
        status:
          type: string
          enum: [planned, cooked, skipped]
        recipeId:
          type: [number, "null"]
        note:
          type: [string, "null"]
    MealPlanUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        title:
          type: string
          minLength: 1
        mealType:
          type: string
          enum: [dinner, lunch, breakfast]
        status:
          type: string
          enum: [planned, cooked, skipped]
        recipeId:
          type: [number, "null"]
        note:
          type: [string, "null"]
    Recipe:
      type: object
      required: [id, name, description, category, prepMinutes, servings, ingredients, steps, favorite]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: [string, "null"]
        category:
          type: string
        prepMinutes:
          type: number
        servings:
          type: number
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        favorite:
          type: boolean
    RecipeInput:
      type: object
      required: [name, category, prepMinutes, servings, ingredients, steps]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: [string, "null"]
        category:
          type: string
        prepMinutes:
          type: number
          minimum: 1
        servings:
          type: number
          minimum: 1
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        favorite:
          type: boolean
    RecipeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: [string, "null"]
        category:
          type: string
        prepMinutes:
          type: number
          minimum: 1
        servings:
          type: number
          minimum: 1
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        favorite:
          type: boolean