# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlan } from './mealPlan';
import type { NurseryStatus } from './nurseryStatus';
import type { ShoppingPreviewItem } from './shoppingPreviewItem';

export interface DashboardSummary {
  dateLabel: string;
  weekLabel: string;
  todayMeal: MealPlan | null;
  weekMenu: MealPlan[];
  shoppingPreview: ShoppingPreviewItem[];
  nurseryStatus: NurseryStatus;
}