# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlanMealType } from './mealPlanMealType';
import type { MealPlanStatus } from './mealPlanStatus';

export interface MealPlan {
  id: number;
  date: Date;
  dayLabel: string;
  title: string;
  mealType: MealPlanMealType;
  status: MealPlanStatus;
  /** @nullable */
  recipeId?: number | null;
  /** @nullable */
  recipeName?: string | null;
  /** @nullable */
  note?: string | null;
}