# lib/api-zod/src/generated/types/nurseryStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NurseryStatusTone } from './nurseryStatusTone';

export interface NurseryStatus {
  label: string;
  detail: string;
  tone: NurseryStatusTone;
}