# lib/api-zod/src/generated/types/recipeUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  /** @minimum 1 */
  prepMinutes?: number;
  /** @minimum 1 */
  servings?: number;
  ingredients?: string[];
  steps?: string[];
  favorite?: boolean;
}